src/requirements.txt
python-dotenv
aiohttp
teams-ai>=1.6.0,<2.0.0
azure-ai-projects 
azure-identity   
botbuilder-core 
botbuilder-schema 
botbuilder-integration-aiohttp 
uvicorn[standard] 